LPS-66316 (SERVICE) Moved non-exported packages to internal package
