+++ modules/apps/mail/mail-service/src/main/java/com/liferay/mail/internal/imap/ConnectionListener.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.mail.imap;
+package com.liferay.mail.internal.imap;
 
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
