+++ modules/apps/mail/mail-service/src/main/java/com/liferay/mail/internal/imap/IMAPAccessor.java
@@ -12,19 +12,19 @@
  * details.
  */
 
-package com.liferay.mail.imap;
+package com.liferay.mail.internal.imap;
 
 import com.liferay.mail.attachment.AttachmentHandler;
 import com.liferay.mail.configuration.MailGroupServiceConfiguration;
 import com.liferay.mail.constants.MailConstants;
 import com.liferay.mail.exception.MailException;
 import com.liferay.mail.exception.NoSuchMessageException;
+import com.liferay.mail.internal.util.HtmlContentUtil;
 import com.liferay.mail.model.Account;
 import com.liferay.mail.model.MailFile;
 import com.liferay.mail.service.AttachmentLocalServiceUtil;
 import com.liferay.mail.service.FolderLocalServiceUtil;
 import com.liferay.mail.service.MessageLocalServiceUtil;
-import com.liferay.mail.util.HtmlContentUtil;
 import com.liferay.petra.mail.InternetAddressUtil;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.log.Log;
