+++ modules/apps/mail/mail-service/src/main/java/com/liferay/mail/internal/imap/IMAPMessageCountListener.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.mail.imap;
+package com.liferay.mail.internal.imap;
 
 import com.liferay.mail.exception.MailException;
 import com.liferay.mail.model.Account;
