+++ modules/apps/mail/mail-service/src/main/java/com/liferay/mail/internal/messaging/MailSynchronizationMessageListener.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.mail.messaging;
+package com.liferay.mail.internal.messaging;
 
 import com.liferay.mail.exception.NoSuchAccountException;
 import com.liferay.mail.mailbox.Mailbox;
