+++ modules/apps/mail/mail-service/src/main/java/com/liferay/mail/internal/model/listener/UserModelListener.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.mail.model.listener;
+package com.liferay.mail.internal.model.listener;
 
 import com.liferay.mail.mailbox.Mailbox;
 import com.liferay.mail.mailbox.MailboxFactoryUtil;
