+++ modules/apps/mail/mail-service/src/main/java/com/liferay/mail/internal/util/HtmlContentUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.mail.util;
+package com.liferay.mail.internal.util;
 
 import com.liferay.portal.kernel.util.StringPool;
