+++ modules/apps/mail/mail-service/src/main/java/com/liferay/mail/mailbox/IMAPMailbox.java
@@ -20,8 +20,9 @@ import com.liferay.mail.constants.MailConstants;
 import com.liferay.mail.exception.MailException;
 import com.liferay.mail.exception.NoSuchFolderException;
 import com.liferay.mail.exception.NoSuchMessageException;
-import com.liferay.mail.imap.IMAPAccessor;
-import com.liferay.mail.imap.IMAPConnection;
+import com.liferay.mail.internal.imap.IMAPAccessor;
+import com.liferay.mail.internal.imap.IMAPConnection;
+import com.liferay.mail.internal.util.AccountLock;
 import com.liferay.mail.model.Account;
 import com.liferay.mail.model.Attachment;
 import com.liferay.mail.model.Folder;
@@ -32,7 +33,6 @@ import com.liferay.mail.service.AccountLocalServiceUtil;
 import com.liferay.mail.service.AttachmentLocalServiceUtil;
 import com.liferay.mail.service.FolderLocalServiceUtil;
 import com.liferay.mail.service.MessageLocalServiceUtil;
-import com.liferay.mail.util.AccountLock;
 import com.liferay.petra.mail.InternetAddressUtil;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.LanguageUtil;
